Updated the migrations with the latest schema and created a seed file.
